Make source optional.
When calling create_indicator_from_tlo a source is not required. It will
always inherit the sources from the initial object, and optionally add
sources if requested.

This means the view does not require it either. This unbreaks services
which were calling into the view via AJAX (whois, shodan, and
dataminer).