Rework add_relationship().
Callers must now pass in the object (rel_item) and rel_id and type_ are
now not supported. This simplifies the logic in add_relationship() and
allows us to make certain promises that allow for more flexibility in
creating relationships.

For example, it is now the case that we will always save rel_item if a
relationship can be made. It is still up to the caller to save the other
object.

We can now do special case relationships that were more difficult before
this change. For example, when relating something to a versioned
backdoor we will attempt to also relate it to the family backdoor also.

If you have any code calling add_relationship() using rel_id and type_
you must adjust accordingly.