Removing CybOX/STIX from core (first stab).
Removes all of the code specific to CybOX/STIX.

The "Download TLO" feature is now JSON-based instead of STIX.

The only code left is vocabulary-related. Once the vocabulary is set we
can remove the rest of the CybOX/STIX stuff (basically in
core/management/commands/create_*).

Along with changing the vocabulary we will most likely change a lot of
the logic that handled Object Types and splitting on '-'.