Fix PATCH API issues.
The derived resource names needed more fixing to be turned into the
correct string representation of the TLO type they represent. Now the
'type_' will be the correct TLO type. Determining the correct action was
also buggy. If the type matched we would try to find the action in the
type's action dict even if the action was in 'Common'. Now if we can't
find it in one place we will properly look for it in the other.

Also fixed a variable re-use in another part of the core API which I
don't believe was causing problems but pylint was complaining.