Building out more vocabulary.
Adjusting the names of some of the files to keep things more organized.
Building out vocabulary for indicators, ips, objects, relationships, kill
chain, confidence, status. Added docstrings where needed.

RelationshipTypes class has a classmethod `inverse()` for finding the
inverse relationship.

Due to the additions of docstrings and the `inverse()` classmethod I had
to rework the `values()` classmethod all of the classes inherit from the
vocab class so it properly builds the values list.