Add in Indicator and Object support.
Indicators and Objects now use the new vocabulary. There is a special
Object type value "File Upload" which will allow you to upload an actual
file to the system like you could previously. Due to laziness I decided
to leave in the "is the file you just uploaded also a PCAP?" detection
and it will still do the Relationship dance. If that winds up breaking
the proper fix should be to just remove that feature as people should
really be using the PCAP TLO for this.

Since the vocabularies are easier to work with, there is now an icon in
the Object listing row template which will tell you if the Object type
is one that is not also an Indicator type and thus you cannot add that
Object as an Indicator.

There was a LOT of cleanup and stuff to do for this with all of the
tangled mess with Domains, IPs, and the Object bulk uploading interface
thing. I suspect some fallout from this.