Removed a huge memory issue
Do file checks when objects are of file datatype and
ensure filenames are of md5sum hashes when looking
to delete files.

Found when indicators are deleted it takes time and eats up
a ton of memory.  The big bottleneck was when Objects of the
Indicators were removed it brute forced file look ups
especially when they weren't of enum type 'file'.
mongoengine looked up and cached all files of the TLO's to ensure
there were no other references before deleting the file.

In cases of LARGE databases this took minutes to delete one
indicator and ate up gigs of memory.