Make it easier to provision services in a Vagrant dev environment.
This assumes the crits_services repository is checked out beside the
crits repo.

Also, some changes were made to not break the entire CRITs app if a
service is missing dependencies. Prior to this change, the Yara, TAXII,
and Snugglefish services would cause Django errors when loading the
CRITs home page.