Add preferred list to Actions.
When creating a new IndicatorAction (this will be a misnomer in the
future, more on that later) you can now select which TLOs this Action is
"preferred" for. An action having "preferred" status means that from the
details page or the listing page you can click a single button and all
actions that are preferred for that TLO type will be applied.

Currently the only one that this matters for is Indicators, but the code
has been built so that when Actions become an attribute of all TLOs the
same interface can be used.