Properly set status for Indicators...again.
If the Indicator is new, set the status to New or to the provided
status. If the Indicator is existing and the status is New, update to
the provided status. If the Indicator is existing and the status is NOT
New but the provided status IS New, do not update the status.