Make Actions unique to specified TLOs.
When the "Add Action" form pops up, it will now query the database for
which Actions are specified for the TLO type in question. This allows
you to have Action Types that will show up for a certain TLO type, but
not others. This also means you need to change all of your existing
Actions to support Indicators otherwise they will not show up in the
list.