Allows for camelcase Indicator values.
We now store the value as-is, but store the lowered value in a "lower"
attribute of the Indicator. When searching for duplicates we will search
against the "lower" field using the value.lower() of the new Indicator.
If a match is found, we will update the existing Indicator as we
normally would on a match, but we will append to the "description" field
what the actual value was that was added. This allows people to store
historical reference for camelcase values that are different but the
same when lowered.

This comes with a schema bump for Indicators so we can backfill the
"lower" field.