Make Indicator Threat and Attack Types a tag list.
These are functionally equivalent to sectors. You can modify the list as
you see fit through the UI and clicking on one will search for all other
Indicators with that type specified.

There is a schema change and an entry in UPDATING outlining the change,
the updating command to run if you wish, how to find Indicator
duplicates in your database for *manual* reconciliation, and the changes
to the PATCH API for developers.