If someone uploads a PCAP as a sample, detect it and DTRT.
This is really rough code as the sample uploading code is a freaking
mess. This will allow for a single PCAP to be uploaded or a set of PCAPs
in a zip, or theoretically a zip containing both samples and PCAPs. It
will detect the PCAP by magic number and upload it as a PCAP instead. In
bulk situations it will return just the MD5 of the PCAP. This makes it
hard to determine in the upload list whether it was a sample or a PCAP,
so for now there are links to both details pages, so if it's a sample or
PCAP one of the links will work.

This also exposed some issues where the good informational messages in
Sample uploading was being masked by a default message. This was fixed
but caused other errors. I hope those were fixed in this.