Making horizontal menu less bad.
Still not great, but this makes the horizontal menu scalable. It will
now horizontally scroll as the list gets bigger. This will help with the
issue on the Analysis tab (specifically with Samples) where so many
services could be enabled that it would start flowing over the first
analysis result making it hard to expand/read/etc.