Handle exception caused by bad data in AnalysisResult
If a service produces bad data in its analysis results such that the data can't be written to the DB, the resulting error is not handled.  This leads to an endless spinning circle in the `Analysis` tab of a TLO as if the task never completes.
This fix assumes that the bad data is most likely in the `results` component and attempts to clear that data only while adding a log entry with the error.  If the DB update still fails (bad data in log entry?), another update it attempted with only the new error log entry.