This fixes an issue with editing/deleting actions.
There were situations with preferred actions where each action added
would get the exact same datetime of when it was added. This caused
issues with editing and deleting since we keyed off of the datetime
only. This change modifies a couple things.

First, we now get a new datetime for each preferred action added.
Second, since that doesn't 100% guarantee uniqueness, we now require the
action type to be passed in as well when deleting. Third, we no longer
allow for editing the action type and it is now used as well as the
datetime to find the action you are editing.

Ran a collectstatic which picked up a couple other changes that were
lingering and never collected before.