Only write Target email_count to DB if value changes
Every time the mapreduce job runs, regardless of whether or not the email_count has changed, CRITs re-writes to the DB each Target where the count is not zero. While somewhat trivial, unnecessary DB write operations are relatively expensive and probably should be avoided.  Additionally, the write operation updates the "modified" timestamp, which could be misleading.