Fixed an issue with removing relationships.
We were running into issues where enumerate was extending past the end
of the list after the matching relationship was removed. Sometimes this
resulted in the addition of a "null" entry in the list due to how we
were removing them.

Now we use ".remove()" instead of "del" and break after it is removed to
stop the enumeration.

Also removed a redundant ".save()" of the right-side of the
relationship.