Fixing load times.
I hope this doesn't break things, but this causes load times to drop
from 17s to around 4s. It also drops over 500 reads from the database
for a single request and over 500 full recalculations of the user's ACL.

If this causes problems, comments are there to note the changes. The
only things I can think of is a bug a long while ago where users were
having bled ACLs but it would seem very odd that this would be the cause
of the issue.