Fix an issue with lingering ACL overwrites.
There was an issue where the last logged in user would overwrite the ACL
for all users who already logged in and were active.

To fix this I decided to write the user's ACL to the database. This will
cause slightly more overhead in pulling the user document out of the
database, but the `acl` attribute will act as the previously-known
"cached" version of the ACL.

To accommodate the need for updating, there's also a new
`acl_needs_update` attribute for the user. This will generally be set to
False, but two conditions will set this to True:

    1) The user's account has been changed in ANY way.
    2) A role the user has been granted has been changed in ANY way.

If those conditions are met, the user's account will get flagged.
Whenever we do a `.has_access_to()` check on that user, if they are
flagged for updating, we will update their ACL first (which will save it
to the database) and then proceed with validating their access.

This seems to have fixed the issue as far as I can tell. This did come
with another issue which was how to process the user's ACL as it is
stored as a dictionary in the database. A few tweaks were made for
iterating over the ACL to find a sub-key as well as looking up their
source access.