Attempt to fix performance problem raised in Issue #913
First run at attempt to push the TLP filters to the database query, instead of
doing it on the server-side. This is especially expensive because it causes
paginated views to fetch all elements from a collection. We have over 2M objects
in some collections, and this made the application unusable, regardless of
hardware, post tlp-addition.

I have left the "old code" in place, but commented out, so that the crew
can compare the intent/logic of the original implementation with my refactor.

Once I get thumbs-up, or tweaks that correct errors I've introduced, I'll
commit the removal of the commented-out code.

It is super-fast now, btw.