Add 'c-field__eq' support to API, and fix de-quote
The 'querydict[field] = {op: val}' operation seems to be failing if the
user requests ?c-value__ne=something, and the fix appears to be
remove_quotes(val). So fixing this.

Additionally, when making an API call with regex=1, there didn't appear
to be a mechanism to force *some* fields to be non-regex searches. Since
this is a potential query optimization (limiting what regex gets applied
to), I added support for c-field__eq, which will use the existing "regex
override" logic to force any specific field(s) to be equality-matched, rather
than regex matched.

Using the documentation
https://docs.mongodb.com/manual/reference/operator/query/eq/#op._S_eq
I simplify {'field': {'$eq': remove_quotes(val)}} to be
{'field': remove_quotes(val)} because the docs say the two are
equivalent anyhow, and this would reduce the data size when querying
mongodb, for larger queries.