Fix case where XHR login reauths an existing session
When XMLHttpRequest attempts to authenticate an already-logged-in
session, the response is rendered as HTML. This causes the "user is
already authenticated" session cookie check to only occur on the /login/
route view when the request is not Ajax. If it is Ajax, the code
proceeds to create a new session and return JSON code. The HTML response
in the current behavior causes anything expecting JSON to generate a
parsing error.