Attempt to chase a bug I introduced with the __ne support
Basically, there are some cases where the value of a field in the query
string is represented as a comma-separated list of values. There is code
present earlier in this function to identify these and do the
tokenize-and-listify operation. However, that renders val as a list,
rather than string, type by the time it gets to the code to populate the
field constraint in the query dict (which is coded to call
remove_quotes(val) ).

So the solution is to test whether val is a list, and if it isn't, run
the remove_quotes(val). Otherwise, just pass the already-decoded list as
the rvalue in the assignment.