Call remove_quotes only if val is a str
In the cases where val was a str type, this would always run the
remove_quotes function. However, if the val was anything other than a
list type (such as a datetime), it would also attempt this translation.
This change limits the translation to only occur on str types, which I
think is the appropriate action.