Update dependencies/requirements
django-debug-toolbar has to be <1.9 due to the more recent versions forcing Django 1.10 as requirement

mongoengine from 0.9 to the latest: 0.15.x makes CRITs painfully slow, so we're going back to 0.8.8 as default. However, the current code should be able to handle the latest versions.

pymongo 2.8.1 works probably the best with 0.8.8, when using the latest mongoengine, try to use the latest pymongo 

celery and django-celery were dropped, since I couldn't find it in crits nor crits_services.